MoeedWeb/requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
exceptiongroup==1.1.0
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytest==7.2.1
pytest-django==4.5.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
